+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] - 2019-05-08
+
+### added
+
+- Support for the AZ module as well as the AzureRM module. #4
+
+### Fixed
+
+- Bugfix to get rid of empty values in Test-ArmDeploymentResource
+- Bugfix for empty errormessages in Get-ARMDeploymentResource
+- Bugfix Test-ARMDeploymentResource: Network security group output is wrong #3
+
 ## [0.2.1] - 2019-04-18
 
 ### added